Include information about third party tools
Other tools (like Babel) are available and can solve problems devs have with the react-tools transform. I wanted to make devs aware of them, and one way to do this is via the wiki. If such a list should instead be included into the site, that's fine for me as well.