Statically check PropTypes.oneOf when provided string literals
Technically, oneOf accepts any values and compares using `===` at
runtime, but most of the time we only care about string literals, for
which it's trivial to create a type in flow.

Note that if any values in the array passed to oneOf is not a string
literal, the inferred type is `any`.