Treat React props like an object literal
When props flow into react element proptypes, they are not treated as
object literals, because their reason is not exactly "object literal."
This causes the props to be unified with the proptypes instead of
compared using subtyping rules.

The real fix to this is to carry around "this is literal" information
separately from the reason string, but until we have that, I'm making
this hack even dirtier.

This has a nice benefit of simplifying type failures, as you can see in
the changed test output. I verified that all removed error messages are
inverses of messages that still exist -- i.e., we have not lost any type
checking.