Change block type to record with scope type information
In order to support let/const, we need to know what kind of scope each
block is.

Currently, the only blocks we have are what I am calling "Hoist" blocks.
Whenever an entry which should be hoisted is added to the env, it should
be added to the closest "Hoist" block.

Let/const identifiers should be added to the nearest "Block" block.

Currently we only create blocks corresponding to hoist points.
Subsequent commits will need to create a great many more blocks to
support block-level scope.

NB: We should probably rename "block" to "scope."