For loops introduce a lexical scope for initialized variables
I extracted variable_declaration from outside of statement_decl. I
changed the name to variable_decl to mirror statement_decl, but would be
happy to revert that.

I don't totally understand the change I made to widen_env, so please
take a close look there. This is very similar to a change I made in
clear_env, where the modification happens in all scopes up to the
closest VarScope.

Once I understand this better, and if it is clear that is the correct
thing to do, that pattern should be extracted out into a function.