error position and trace improvements
Summary: This diff adds the machinery to track operational
positions - positions in the code of the operations
which provoke certain typechecks, regardless of the
positions of the types being checked.

These positions must be recorded proactively by pushing
and popping within the flow algorithm. Given the fragility
of that code, and the churn in test files that needs to
be checked when positions change, only kind of operation
has actually been adapted to use them so far: function calls.
Later diffs will carefully add more kinds of operations.

Also in this diff - some improvements to trace printing.
This is WIP, traces will become much more informative with
some more work. But they're a bit better now.

Reviewed By: @avikchaudhuri

Differential Revision: D2047246