error message formatting
Summary: The old message format was (brackets mean optional):
```
[<position 1>: <reason 1>]
<conflict description>
  <position 2>: <reason 2>
```
Now that messages are getting more parts and more variation,
we need some additional markers. This diff uses this format:

```
[<operational position>: <operational reason>
Error:]
[<position 1>: <reason 1>]
<conflict description>
<position 2>: <reason 2>
```

So e.g. instead of

```
/data/users/bhosmer/scratch/foo.js:7:12,17: null
This type is incompatible with
  /data/users/bhosmer/scratch/foo.js:4:25,30: string
```

we print

```
/data/users/bhosmer/scratch/foo.js:14:5,20: function call
Error:
/data/users/bhosmer/scratch/foo.js:7:12,17: null
This type is incompatible with
/data/users/bhosmer/scratch/foo.js:4:25,30: string
```

Note: in cases where the operational position/reason is the same
as position 1/reason 1, we don't print the new stuff. There may
be better ways to eliminate the redundancy. Also, operational
reasons aren't yet being recorded for all ops. WIP

Reviewed By: @avikchaudhuri

Differential Revision: D2050771