Move handling of "this" type into Typing_phase.localize
Summary: The "this" type is handled in a number of places through out the type checker.
Now that the Phase refactor is in, I worked to centralize this logic the best I
could. I introduced a new ##decl ty## ##Tthis##. This avoids doing
instantiation of ##this## at all in the decl phase. When translating to the
locl phase we instantiate ##this## as we did before this diff. To do this
Typing_phase.localize must be supplied with what the ##this## type
should be and the set of substitutions that are necessary.

I also refactored how instantiation works in the locl phase as well since it is
so tightly coupled with the ##this## type. Now instantiation is handled by
localize for most cases.

This has resulted in some perf savings from what I can tell of a second or two
off both the Typing_decl and Typing phases. With a bit more refactoring I can
eliminate all instantiation that happen during the locl phase done outside of
localize.

Reviewed By: @elgenie

Differential Revision: D2034276