Disentangle PropTypes from ReactComponent's P
Summary: Having `propTypes` be a `$Record<P>` meant that you needed an entry
for every property that Flow knows about. Since the react team wants to move
away from the runtime validation of PropTypes, we don't want to enforce this.

We still want to keep `propTypes` somewhat in sync with `P`, and `$SuperType` should do this.

Reviewed By: @bhosmer

Differential Revision: D2053088