Speedup up build by only building hack binaries if any source file is modified
Summary: Building hack takes about 7s on my machine ("time make -C hphp/hack") even if no files are changed. This can be dramatically reduced by not invoking a useless recompilation of hack if no files are changed. This patch reduced the build of hack, if unchanged, from 7s to 0.4s. On my machine it also reduces the build time of the complete hhvm tree by 20%
Closes https://github.com/facebook/hhvm/pull/5318

Reviewed By: @jwatzman

Differential Revision: D2058421