constructor call fix
Summary: Previously, constructor calls wouldn't work on classes where the instance type
was not resolved. Also, the logic that worked repeated parts of method call
logic, and didn't handle the case where a constructor method was not found.

This diff simplifies the logic of constructor calls to reflect runtime behavior.

Reviewed By: @bhosmer

Differential Revision: D2063340