show parse errors from unchecked but required
Summary: Flow's policy has been to suppress parse errors in
non-@flow files. This leads to confusion when such
a file is imported into a flow file: the user gets
only a module-not-found error, which has notoriously
many possible causes. With this diff we track imports
of unparsed files, and report parse errors for files
that have been imported, @​flow or not.

Note: to track Haste imports without parsing, we have
to grep for @​providesModule rather than parsing it
from the comment header. At the limit our grep will
accept annotations that the parser would reject, but
this should be okay since it's always fixable by
fixing the parse error.

Code is ready for review but need to add tests for
2^4 cases: parsed * imported * @​providesModule * @​flow.

Reviewed By: @gabelevi

Differential Revision: D2060629