fix get-def
Summary: Repeated calls to get-def would cause exceptions to be thrown.
Ultimate cause of this was a failure to properly merge required
contexts, due to a failure to properly clear a global cache
before merging. Since the lifetime of the cache is a single
merge, here we just parameterize the merge with a cache object,
eliminating one small piece of global state. One step at a time.

Reviewed By: @avikchaudhuri

Differential Revision: D2067825