only allow strings and numbers in computed properties
Summary: computed properties were extremely loose before. if you pass a string and we know the literal value, we verify that //that// property exists; if the object has an indexer, then we verify the key is compatible; otherwise, we allow anything.

this diff changes the last case, so that we only allow strings and numbers. this notably bans `null` and `undefined`, even though JS allows them, because most likely you didn't want a bunch of stuff all colliding at `obj["undefined"]`.

future work: we allow any number (as we did before), but ideally we would cast the number to a string and then verify it the same way we do strings (see the two TODOs in the unit test)

Reviewed By: @gabelevi

Differential Revision: D2070593