re-instate Haste_module_preprocessor for now
Summary: I pulled this out in the `module.name_mapper` diff because it's superceded by the new config option, but it's useful to be able to test existing codebases with the master version of Flow without having to patch in the new config first.

This just adds the preprocessor back, and we can rip it out again after the next deploy.

Reviewed By: @gabelevi

Differential Revision: D2072815