fix get-def
Summary: 1. Fixed get-def

The get-def command uses Flow_js.resolve_type to crack open
tvars (as when the definition being sought comes from an
imported or required module).

Flow_js.resolve_type relied on the unspoken rule that the lower
bounds of a tvar were fully resolved concrete types. However
at the moment instances of the CJSExportDefaultT wrapper may
appear in lower bounds in certain situations when they haven't
been intercepted by an unwrapping import operation.

CJSExportDefaultT will disappear or change behavior in an
upcoming refactor of imports, so for now we just have a shim,
typemap_keys, that knows how to unwrap TypeMap.keys. It should
be removed when the source problem is fixed.

2. Fixed get-def some more

*Always type-annotate variables that hold unmarshaled results!*
There was a classic unmarshal-to-different-type bug that caused
file paths in non-JSON positions to print as garbage. Caught and
fixed by adding a type annotation on line 70 of getDefCommand.ml.

3. Added non-check commands to runtests.sh

This is very crude for now, but it's good enough to enable
tests on flow client commands that autostart a server, like
this one.

4. Added an inaugural get-def test. :)

Reviewed By: @avikchaudhuri

Differential Revision: D2070608