flow + unification cleanup
Summary: This diff refactors the data structures and algorithms for flow and unification
so that these systems can co-exist harmoniously.

1. The representation of type variables and their constraints in the graph is
   now more structured.
2. The steps involved in the propagation of constraints have been broken down
   and reused.

I've added some commentary on the invariants maintained by the implementation.
Given the combinatorial nature of cases, it is possible that I've missed listing
some invariants; I'll keep adding them in later passes, but as it stands this
diff is ready for review.

Note that some dead code (e.g. gc) is known to be broken in the new setup.
I'll clean those up in later diffs now that we know exactly how type variables
behave under flow + unification.

Reviewed By: @bhosmer

Differential Revision: D2056317