Factor out client connection code
Summary: clientCheck and clientBuild duplicated a lot of the connection logic.
This commit factors that out. It also replaces the use of exceptions in
favor of the Result type, which is exhaustively checked.

I was too lazy to do the same cleanup for Flow, so I moved over the
`ClientException` and `ClientUtils` stuff that it still needed.

Reviewed By: @dlreeves

Differential Revision: D2074828