Install member expression refinments in the appropriate scope
Member expression are refined by creating a special "$REFI x.y" name and saving
that into the env. Introducing LexicalScope and VarScope scope kinds to the env
introduces some extra complexity here, as we need to be careful about the
lifetime of the refinement.

Ultimately, the refinement should be "installed" in the same scope where the
original variable was defined. Otherwise the refinement could "outlive" the
variable itself or be discounted too soon (see tests).

This only matters when a refinement outlives the conditional section where it
was introduced, which can happen after an abnormal exit. No abnormal exit can
cause a refinement to break out from the function where it was introduced, so we
terminate the search for the parent object once we hit a function scope.