Structural subtype all the things ~> interfaces
Summary: Currently we do structural subtyping when you flow an

```
ObjT ~> InstanceT
```

So you can use an Object as a class instance or interface instance if it looks
right (we might want to rethink using an object as a class instance due to
nominal subtyping).

This change allow anything to be an instance of an interface as long as it
looks right.

Reviewed By: @bhosmer

Differential Revision: D2077992