add dump-types command
Summary: adds `flow dump-types file [--json]` command which is similar to `type-at-pos` but it lists the type of every range in the file.

i think this could be useful to build a tool that shows type information inline (e.g. diffusion integration) or as a basis for a `--color` option like hack has.

mostly, though, I just wanted to figure out how to write a command that queries some info from the flow server.

Reviewed By: @gabelevi

Differential Revision: D1877334