Env_js.set_var should not create new entries
`update_scope` is only ever called as a result from `set_var`, which is
called in a variety of contexts in `Type_inference_js`.

I got here by adding a `scope_kind` field to the `scope_entry` record
type, in order to support TDZ semantics for lexical declarations.
Originally I thought I would need to thread a `scope_kind` through these
methods, but it turns out that this default isn't really necessary.

By my analysis, `set_var` is only called in the type inferencer's
`statement` pass, after the `toplevels` pass has already run. The
`toplevels` should have called `init_env` for any introduced identifier,
so `set_var` should never be invoked with an entry that isn't already in
`env`. In fact, if that does happen, it's an error.

I could be wrong about that, but all the tests pass.