Add scope_kind field to scope_entry record
This is necessary in order to implement TDZ error checking. Only
lexically-scoped variables are subject to TDZ, so we need to store the
scope kind along with the entry itself for lookup later. Until now, we
only used the scope kind in `init_env` to place the entry in the correct
scope.

In future commits, I will remove the now redundant scope_kind parameter
to init_env. I will also add a simple TDZ check.