Block-scoped variables still hoist
Even though the block-scoped variables foo and bar are defined in the
above scope, within the inner scope, foo and bar are still uninitialized
because the variables still hoist.

I expected this to work already, but due to a logic bug in exists_in_env
we didn't catch this case. exists_in_env takes a scope_kind argument,
which controls how far up the env we need to look: a VarScope will look
up to the enclosing VarScope, and a LexicalScope will only look at the
head scope.

However, the previous implementation only returned scopes that match the
passed scope kind exactly, so an entry in a LexicalScope would never be
returned when the search scope_kind was VarScope.