Return scope along with entry from exists_in_env
In the event that we init_env with a variable that is already in scope,
we need to know what kind of scope to catch errors declaring
block-scoped variables.

This commit just changes the return value from exists_in_env. Future
commits will use this information to implement the necessary checks.