Switch cases are surrounded by a fresh lexical scope
Introducing the same let-bound variable between two cases is the same as
shadowing a let-bound variable in a single lexical scope.

The AST for switch actually has a "lexical" flag, which is always false
and marked (* TODO *). The SpiderMonkey AST docs say that the flag is
only set if lookahead determines there is a let-binding in the switch
statement.

However, I don't think that matters for the purposes of type checking. I
think it's safe to always introduce a lexical scope here.