Warn on unsupported class members
Summary: Before this change, flow would silently ignore these properties and
instead users would see "Property not found" errors when they went to
use the desired properties.

I intentionally didn't include a fallthrough case, because the compiler
will warn if this pattern ever becomes inexhaustive in the future.
However, that's currently just a warning. Can we make it an error?

Also, because of the way the AST is defined, some of these cases are
actually syntax errors, like computed- and literal-key method bodies. I
would be happy to also make the changes to the AST to restrict the
representation to exclude impossible cases, if that seems important.

See #444
Closes https://github.com/facebook/flow/pull/461

Reviewed By: @jeffmo

Differential Revision: D2097017

Pulled By: @mroch