Add test illustrating unification error in current code
Because flow tries to unify these two types, users see errors going both
ways. Unification isn't necessary here, because we know the right-hand
side is literal.

Next commit fixes this. This commit is for the benefit of the reader,
who wouldn't otherwise see any expectation change in the impending
commit.