refine globals
Summary: we were skipping refinements on globals. I'm not sure why... I tried to blame it, but couldn't go back that far.

I see no reason not to refine globals... maybe it was a perf issue before, if we were sharing globals across files and accumulating tons of upper bounds?

Reviewed By: @bhosmer

Differential Revision: D2101295