Fuzzy search service for flow
Summary: Plugged flow into the Hack symbol search service (previously monocle).
Usage:
$ flow search my_function_name

To very quickly find all the matching symbols in the codebase.
The indexes are maintained by the server and always match the current revision.
It only works with classes and functions for now. It's not ideal because JS has a module system that
in this case is completely ignored. But I figured it would be nice to start with something minimal and
revisit to make it better (more contextual).

Reviewed By: @bhosmer

Differential Revision: D1983816