Factor out class AST
Before this change, the AST had two separate represenations for class
statements and class expressions (although expressions re-used most of
the statement AST).

I want to write a function that creates a class type for a class
regardless of its statement/expression status, so I factored out the AST
into a module outside of Statement and Expression.

I didn't touch the dts folder. I'm not really sure what that's for, but
I can probably make a similar change there if requested.