typeof annotations
Summary: Diff adds support for type annotations of the form `typeof x`, where `x` is the name
of an in-scope variable. See tests/typeof for examples.

The implementation is half-generalized. It uses a new BecomeT type, which is a general-
purpose tvar latch, not restricted to use here (It's used in a couple of other places
in this diff to demonstrate.) OTOH type annotations in general will go through further
generalization shortly, both to rationalize and to fix the leakage problem.

Reviewed By: @avikchaudhuri

Differential Revision: D2110348