[Parser] Add infinite lookahead
Summary: This change adds the ability to get an arbitrary number of tokens of lookahead
from the parser.

I left the single lookahead as is, because so much relies on there being exactly
one token of lookahead, and on the fact that the lex buffer and lex_env are in
the state produced by lexing one token of lookahead. After trying to fix that
for a while, I decided to implement lookahead completely separately. Hopefully
this will let us reduce our reliance on weird things like vomit and advance, at
which point removing those assumptions may be easier.

Reviewed By: @gabelevi

Differential Revision: D2092764