recursion limiter
Summary: Add a circuit breaker to __flow that prevents it
from recursing beyond a predefined depth
(currently 100).

When this limit is reached, a "recursion limit exceeded"
error is added, with as much trace info as we have (a single
level if traces are set to 0, the full trace otherwise).
Subsequent typechecking proceeds as usual.

Note: this diff does not remove the type pair cache -
that'll come in a subsequent diff from Avik - but
everything should be in place to trap unforeseen
recursion once that diff has landed.

Reviewed By: @avikchaudhuri

Differential Revision: D2117093