(null|undefined) + number === number
Summary: per ES6 12.7.3.1, `null` is cast to +0 and `undefined` is cast to `NaN` when part of an addition expression.

arguably, we should error if you do `null + number` or `undefined + number`... we can certainly have that debate here or later. but we definitely should not be treating the result as a string like we do now.

Reviewed By: @avikchaudhuri

Differential Revision: D2082341