Do less funky index work in heap_refinement
I think this construction makes it clear what we're doing. One
interesting bit, if a refinement is for an internal name, like `this`,
then the key will be `$REFI .this`, so the returned obj key will be the
empty string.

We won't find a binding named "" in `install_refinement`, so we'll end
up installing the refinement in the nearest VarScope, which is the
correct behavior.

Existing tests fail if we get this wrong, so I feel safe leaving things
this way.