Replace for_type flag on scope entries with new binding type
This does not change the for_type argument on lookup functions, which is
not exactly clear to me. I need to disentangle the call stacks that
depend on that variable to figure out the intent, then create more
specific APIs in Env_js to accomodate those use cases explicitly.

TypeBindings have the same scoping rules as VarBindings, which is how
they have been used so far, but it might not be the case in general. If
it does turn out that some type bindings should follow block scoping
rules, things could get complicated.