[easy] remove unsupported --version flag from a bunch of commands
Summary: `--version` is only actually handled by the status/default commands, but lots of other commands accept it and do nothing.

this moves it to just the default command, not even the status command (just `flow --version`). this matches other commands with subcommands that I could find, like `hg` and `git`.

Reviewed By: @gabelevi

Differential Revision: D2125389