Ways to suppress flow errors
Summary: The general idea is to enable someone to suppress a Flow error on line
N+1 by adding a special comment on line N. We basically need to keep track of
all the suppression comments in all files and then filter the errors when we're
done merging.

To be clear, if an error mentions 3 locations A, B, and C, you can add an error suppressing comment on the line before A, B, or C to suppress it.

Reviewed By: @bhosmer

Differential Revision: D2097222