[rfc|flow] Add StyleSheet and overloads for createElement
Summary: This adds overloads for `document.createElement` so that flow knows what to
expect. It also adds `HTMLStyleElement` and `StyleSheet`, as per

https://developer.mozilla.org/en-US/docs/Web/API/HTMLStyleElement
https://developer.mozilla.org/en-US/docs/Web/API/StyleSheet

I know @weicool wanted something like this in D2124593, and I think this will get
us partway. However, the spec says that the `sheet` property is a `StyleSheet`,
whereas @weicool actually needs the "subclass" `CSSStyleSheet`. I'm not sure how to
resolve that part.

Reviewed By: @avikchaudhuri

Differential Revision: D2126133