arrow fix
Summary: A recent change to how arrows behave accidentally resulted in allowing
`this` in their types to be rebound based on context. In particular, calling
them as functions would re-bind `this` to the global object, causing errors:
"<original `this`> is incompatible with global object."

This diff restores the bound-`this` behavior that arrows share with methods,
while keeping all the other goodies in the recent change.

Reviewed By: @gabelevi

Differential Revision: D2123421