stop using hack errors
Summary: since we don't use error codes, hack errors weren't buying us anything and were adding a confusing layer of indirection.

I do think that using `Pos.t` instead of `Reason_js.reason` in the errors (as hack does) makes more sense since we never use the reason for anything other than the pos, so I'm going to go through and try to change to that, but that'll make it even less valuable to pull in hack's `Error`.

Reviewed By: @bhosmer

Differential Revision: D2115656