nuke RecordT
Summary: $Record types were experimental, and kind of rotting as a third object-like
def type besides ObjT and InstanceT.

It turns out that its functionality can be covered by $Enum and dict types.
So there is no use for it.

In fact, using $Enum and dict types, you can get better $Record types because
you can also specify the value type!

Also simplified type of propTypes to Object.
With this diff we could replace it with $Subtype<{[key: $Enum<P>]: any]}>.
But in a follow-up diff, we're going to make propTypes allow spreads etc.
which may interfere with such a type. Moving forward ES6 React doesn't have a
good story around propTypes anyway, and for legacy React that annotation is
not even touched, so Object should really be enough.

Reviewed By: @mroch

Differential Revision: D2132989