fix unification of dictionaries
Summary: unifying objects didn't properly consider dictionaries. now we (a) ensure that the two dicts are compatible, and (b) ensure that any props in one object but not in the other is compatible with the other's indexer.

see the new unit tests. the things in `compatible.js` failed before, and the new things in `incompatible.js` passed before.

Reviewed By: @avikchaudhuri

Differential Revision: D2133072