Export default class
Summary: This change implements the class expression case for the type inferencer, which is very similar to the existing code to handle class declarations. I refactored the AST to enable more sharing of this code. See individual commits for details.

Notably, I didn't touch the `dts` subdirectory, because I am not sure what it is for.
Closes https://github.com/facebook/flow/pull/485

Reviewed By: @avikchaudhuri

Differential Revision: D2139484

Pulled By: @jeffmo