[hack] Move filling of type holes to Typing_phase.localize
Summary: Typing_taccess.fill_type_hole is the last place where we do instantiation on a
locl ty. This function does two things, first creates a new expression
dependent type based on the static class id we are given, second it
instantiates as special generic called a "type hole". This type hole is created
for type constants accessed via ##this##, i.e ##this::T##.

I made a new function make_expr_dependent_ty that we use create the right
expected. This also unifies the concept of ##this## and ##this::T##. The last is
to improve the analysis we do for expression dependent types. Something I'm
working on

Reviewed By: @jwatzman

Differential Revision: D2143871