Clarify assignment_lhs
Summary: Maybe this is obvious to others, but I was confused that it seemed flow
would allow arbitrary expressions as the LHS of an assignment.

I was confused by this code when I went to add a check for mutating
operations to const-declared variables. Eventually I realized that the
parser prevents arbitrary expressions from reaching this code path, but
this code would have saved me some time.
Closes https://github.com/facebook/flow/pull/480

Reviewed By: @gabelevi

Differential Revision: D2115805

Pulled By: @mroch