[hack] Format: Use space more effectively
Summary: The old code was deducting -1 from the right margin each time
we called stmt_list, ostensibly to account for the trailing semicolons.
However, this isn't correct since nested statements do not double up on
their semicolons. This commit fixes it.

We also take into account the trailing `) {` in an `(if|for|while) (...)
{` statement, which we were not doing before. This became more obvious
once the aforemention bug got fixed, as it meant that the right margin
was actually set at 80 instead of 78 in method bodies.

This also revealed that we were handling tuple() differently from
shape() and array(). This has been fixed.

Reviewed By: @jwatzman

Differential Revision: D2144434