Enums converted to classes with static members
Summary: Converted typescript Enums to Flow classes with enum members mapped to static memebers of the class of type "number".

Possible TODO: TS Enums have computation members, whose "values" are determined by the other members. We could just strip off this value. But since this has a meaning, it might be better to just throw an error that we do not supoort this.

Reviewed By: @jeffmo

Differential Revision: D2142698