[hack] Fix JSON escaping
Summary: This JSON library was printing strings using OCaml's escaping rules, which were
not equivalent to those for JSON. This escapes only \ and ", the only characters
that are disallowed in JSON strings as per json.org.

target-determinator-fbcode (whatever that is) fails with a Python exception inside Buck. If that's somehow my fault it needs to be made a lot clearer.

Reviewed By: @bolinfest

Differential Revision: D2138027