Fix `export default class Foo {}; new Foo()`
Summary: Fixes the issue over in https://github.com/facebook/flow/issues/511 where it was reported that `export default` with a named class decl doesn't add the class to the local scope.

Turns out the issue was that we weren't parsing the class as a declaration (but as an expression) -- therefore we weren't declaring it in the scope. This diff just fixes the parser to parse this correctly.

Note that this diff looks much scarier that it is. It's really making 2 code changes:

1) Re-ordering sub-modules in the parser_flow.ml file (this results in most of the red/green)
2) Adding a case to look for `T_CLASS` when parsing `export default` and hopping over to `Object.class_declaration` when hit

Fixed #511

Reviewed By: @mroch

Differential Revision: D2152216