Child imports everything from parent(s)
Summary: Entree´:
This adds another functionality to nested module. Child modules import everything declarared in parent (in fact all ancestor).
This is done so that child modules can access identifiers defined in one of their parent modules even when they are flattened
for Flow declarations.

Main Idea: Push a list of things to be maintained in the dfs stack along with each module node.

Sides:
1. Improved presentation logic of the generated flow declaration files. But still a few lint issues are raised.
2. Added generate_exp.sh file which refreshed .exp files in tests for the new diff.

Reviewed By: @avikchaudhuri

Differential Revision: D2157545