Store module exports in a new specialized type (`ModuleT`)
Summary: This diff refactors the way we store the exports for a module.

Previously each module had a single export type. As inference runs, the export type is mutated or overwritten -- and then at the end of inference we store that type. This strategy was problematic for at least 2 reasons:

(1) In order for both CommonJS and ES modules to co-exist, we had to define the behavior of ES modules in terms of the CJS exports object. This is actually not a big deal for most cases, but it does mean that any ES module features we wish to add have to ultimately reduce to a model that can fit with the CommonJS means of operation (i.e. using `export type` in the same file as `module.exports = Foo` doesn't work).

(2) It also posed problems with the plumbing by which `import/export type` could use to send types between modules. Before this diff, `export type` would have to store type-aliases as properties on the exports object. This dovetails with the previously mentioned problem (clobbered exports + export type don't go well), but it also means that it's possible to convert a type alias into a runtime type by way of `require('MyTypeAliases').SomeAlias`. We can't allow this because it embeds non-runtime types into runtime data -- which would be impossible to strip away.

To address these issues, this diff creates a new type: `ModuleT` which abstracts the notion of exports away from CommonJS exports objects and ES ModuleNamespaceObjects so that exports can be properly normalized without consideration for the way we generally treat objects in the system.

Note that this diff does not yet separate value-types from alias-types in the `ModuleT` structure in order to keep the diff some what incremental (so the aforementioned problem of `require('TypeAliases').SomeAlias` is still possible). I'll send out a follow-up diff in a bit to add this in on top of these changes

Reviewed By: @avikchaudhuri

Differential Revision: D2096461