[hack] Anonymous shapes
Summary: Looking at original HHVM diff (D708631 by @jdelong) that introduced shapes, anonymous shapes were always a goal - runtime support is there since April 2013 (!), but Hack never allowed them. I didn't find the original Hack diff  but in D2019698 @jwatzman remembers @pikatchu being concerned with "What do the error messages look like when we have nested shapes". I've spent some time surveying the error messages for shapes and came to a conclusion that they seem OK, and in cases where they they suck, they do so regardless of anonymous shapes or not, so we can as well ship it.

Here are the errors I looked at:
{P19876888}
Same errors when using anonymous shapes:
{P19876891}
And similar scenario using objects and properties:
{P19876895}

The errors for "value type mismatch" are of same quality in all cases:

  shapes.php:5:15,16: Invalid argument (Typing[4110])
    shapes.php:20:22,27: This is a string
    shapes/s3.php:4:17,19: It is incompatible with an int

  props.php:5:15,16: Invalid argument (Typing[4110])
    shapes.php:20:22,27: This is a string
    shapes/p3.php:4:10,12: It is incompatible with an int

  anonymous.php:13:15,16: Invalid argument (Typing[4110])
    anonymous.php:29:22,27: This is a string
    anonymous.php:6:21,23: It is incompatible with an int

They all point out to (1), integer type annotation with no attempts of reporting the path through the nested structure. None of them mention s1, p1 or point out the outermost shape.

The error for "field missing" is not relevant to objects, and for shapes it also looks the same:

  shapes.php:13:10,14: Invalid return type (Typing[4057])
    shapes.php:15:14,18: The field 'z' is missing
    shapes/s2.php:4:17,18: The field 'z' is defined

  anonymous.php:21:10,14: Invalid return type (Typing[4057])
    anonymous.php:23:14,18: The field 'z' is missing
    anonymous.php:5:19,23: The field 'z' is defined

They point to (2), word "shape" preceding the definition of 'z'.

Finally, the "field undefined":

  anonymous.php:17:22,34: The field not_defined is undefined (Typing[4108])

  shapes.php:9:22,34: The field not_defined is undefined (Typing[4108])

  props.php:9:19,29: Could not find member not_defined in an object of type P3 (Typing[4053])
    shapes/p2.php:4:10,11: This is why I think it is an object of type P3
    shapes/p3.php:3:7,8: Declaration of P3 is here

The error is really bad... in both cases. In case of objects it points out the definition of p3 as well as type on p2 mentioning p1.
Ideally, shapes error would also at least point out the innermost shape word that is actually missing the field (2). It doesn't do this now, but looking at other errors, it's unrelated to having anonymous shapes.

So, to sum up: anonymous shapes don't seem to hurt our error messages. The error messages are (or could be) well working for up to +1 level of nesting. If you want to nest deeper and don't split your code into functions that operate on limited subtrees of structure, you are going to have a bad time regardless if you use aliased shapes, anonymous shapes or even objects.

Reviewed By: @dlreeves

Differential Revision: D2154830